chore(naming): rename RuleName entity to Rule
 - rename entity to reflect its purpose: store rules, not just rule
names.
 - rename Controller and Repository accordingly.
 - update all references in the codebase.
