+++ trading-app-demo/src/main/java/com/pcs/tradingapp/domain/Rule.java
