+++ trading-app-demo/src/main/java/com/pcs/tradingapp/repositories/RuleRepository.java
@@ -1,0 +1,8 @@
+package com.pcs.tradingapp.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.pcs.tradingapp.domain.Rule;
+
+public interface RuleRepository extends JpaRepository<Rule, Integer> {
+}
